# source-main/AgentOS-Phase2-Main/ControlPlane/bl01-08/BL07_POLICY_PLANE_OPENAPI_v1.yaml
openapi: 3.0.3
info:
  title: ARZIN Policy Plane Service
  version: "1.0"
  description: >
    BL-07 Control Plane service. Stores versioned policies and publishes active versions
    for scoped enforcement by Interceptor/SDK. Phase 1 allows observe/recommend only.
servers:
  - url: http://policy-plane:8081
paths:
...
  /v1/policies:
    post:
      summary: Create policy (draft)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PolicyCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Policy" }
        "400": { description: Validation error }
    get:
      summary: List policies
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: policy_id
          schema: { type: string }
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Policy" }
...
components:
  schemas:
    PolicyStatus:
      type: string
      enum: [draft, active, deprecated]
    DecisionClass:
      type: string
      enum: [observe, recommend]
    TopicRule:
      type: object
      required: [topic]
      properties:
        topic: { type: string }
        pattern:
          type: string
          enum: [literal, prefix]
          default: literal
    PolicyScope:
      type: object
      required: [agent_class, channel_id, topics]
      properties:
        agent_class: { type: string }
        agent_id: { type: string, nullable: true }
        channel_id: { type: string }
        topics:
          type: array
          items: { $ref: "#/components/schemas/TopicRule" }
    PolicyConstraints:
      type: object
      required: [decision_classes, signature_required]
      properties:
        decision_classes:
          type: array
          items: { $ref: "#/components/schemas/DecisionClass" }
        signature_required: { type: boolean }
        max_payload_bytes: { type: integer, default: 1048576 }
        ttl_seconds_min: { type: integer, default: 1 }
        ttl_seconds_max: { type: integer, default: 300 }
    PolicyCreate:
      type: object
      required: [policy_id, version, scope, constraints]
      properties:
        policy_id: { type: string }
        version: { type: integer }
        status: { $ref: "#/components/schemas/PolicyStatus", default: draft }
        scope: { $ref: "#/components/schemas/PolicyScope" }
        constraints: { $ref: "#/components/schemas/PolicyConstraints" }
        metadata:
          type: object
          additionalProperties: true
    Policy:
      type: object
      required: [policy_id, version, status, scope, constraints, created_at]
      properties:
        policy_id: { type: string }
        version: { type: integer }
        status: { $ref: "#/components/schemas/PolicyStatus" }
        scope: { $ref: "#/components/schemas/PolicyScope" }
        constraints: { $ref: "#/components/schemas/PolicyConstraints" }
        metadata:
          type: object
          additionalProperties: true
        created_at: { type: string, format: date-time }
...